delete OpaqueImageView2 component, default to OpaqueImageView
