--- src/app/Components/OpaqueImageView2/index.ts
@@ -1,1 +1,0 @@
-export * from "./OpaqueImageView2"
